refactor: rename docs.config.tsx to content.config.tsx
