+++ content.config.tsx
